Add ble-PTM215B-button.js script example (#50)
The PTM215B is a BLE battery less switch that harvest it's energy from the push on it. It can control up to 4 devices with distinct press and release actions, however this is currently not implemented. The release actions is used to improve reliability of that script mainly because it is only possibly to scan 1/4 of the time.

The PR is to share that script to the community but also to open a discussion with Shelly developer on BLE listen time.

Have a nice day.